Add Pomodoro example for how to design for testability.
Also fill in sections describing situations when testing is not very valuable
and when it is. Still need to fill out some of these sections, as well as final
remarks. Committing what I have so far because it was quite a while since I last
committed and I don't want to lose any of this.